Culprit Catcher & Emailing system
Modified the code so that an email address can be set into the config.py
file. If configured and someone other than the box owner is detected the
box will snap a photo of the persons face and email it to the box owner.
This extra security allows the owner to see who else is trying to get
into the box.